Code cleanup.
- Javadoc errors.
- Unnecessary casts.
- Unnecessary else.
- Removed FIXMEs.

+++ src/main/java/com/goodworkalan/glob/Char.java
@@ -57,6 +57,6 @@ final class Char extends Part {
      */
     @Override
     public int hashCode() {
-        return (int) ch;
+        return ch;
     }
 }
